+++ packages/winnow/src/test/aggregate.js
@@ -3,9 +3,9 @@ const test = require('tape')
 const fs = require('fs')
 const winnow = require('../')
 const path = require('path')
-const features = JSON.parse(fs.readFileSync(path.join(__dirname, 'fixtures', `trees.min.geojson`)))
+const features = JSON.parse(fs.readFileSync(path.join(__dirname, 'fixtures', 'trees.geojson')))
 
-test('Get a sum', t => {
+test('Get a sum', (t) => {
   t.plan(1)
   const options = {
     aggregates: [{
@@ -17,7 +17,7 @@ test('Get a sum', t => {
   t.equal(results.sum_Trunk_Diameter, 850305)
 })
 
-test('Get a named aggregate', t => {
+test('Get a named aggregate', (t) => {
   t.plan(1)
   const options = {
     aggregates: [{
@@ -30,7 +30,7 @@ test('Get a named aggregate', t => {
   t.equal(results.total_diameter, 850305)
 })
 
-test('Get an aggregate with a where clause', t => {
+test('Get an aggregate with a where clause', (t) => {
   t.plan(1)
   const options = {
     aggregates: [{
@@ -38,13 +38,13 @@ test('Get an aggregate with a where clause', t => {
       field: 'Trunk_Diameter',
       name: 'total_diameter'
     }],
-    where: `Trunk_Diameter > 10`
+    where: 'Trunk_Diameter > 10'
   }
   const results = winnow.query(features, options)
   t.equal(results.total_diameter, 735026)
 })
 
-test('Get multiple aggregates', t => {
+test('Get multiple aggregates', (t) => {
   t.plan(2)
   const options = {
     aggregates: [
@@ -65,7 +65,49 @@ test('Get multiple aggregates', t => {
   t.equal(results.max_diameter, 130)
 })
 
-test('Get multiple aggregates with a where clause', t => {
+test('Get multiple aggregates specified in the esri way', (t) => {
+  t.plan(2)
+  const options = {
+    outStatistics: [
+      {
+        statisticType: 'sum',
+        onStatisticField: 'Trunk_Diameter',
+        outStatisticFieldName: 'total_diameter'
+      },
+      {
+        statisticType: 'max',
+        onStatisticField: 'Trunk_Diameter',
+        outStatisticFieldName: 'max_diameter'
+      }
+    ]
+  }
+  const results = winnow.query(features, options)
+  t.equal(results.total_diameter, 850305)
+  t.equal(results.max_diameter, 130)
+})
+
+test('Get multiple aggregates specified in the esri way when the input is a string', (t) => {
+  t.plan(2)
+  const options = {
+    outStatistics: JSON.stringify([
+      {
+        statisticType: 'sum',
+        onStatisticField: 'Trunk_Diameter',
+        outStatisticFieldName: 'total_diameter'
+      },
+      {
+        statisticType: 'max',
+        onStatisticField: 'Trunk_Diameter',
+        outStatisticFieldName: 'max_diameter'
+      }
+    ])
+  }
+  const results = winnow.query(features, options)
+  t.equal(results.total_diameter, 850305)
+  t.equal(results.max_diameter, 130)
+})
+
+test('Get multiple aggregates with a where clause', (t) => {
   t.plan(2)
   const options = {
     aggregates: [
@@ -80,7 +122,7 @@ test('Get multiple aggregates with a where clause', t => {
         name: 'max_diameter'
       }
     ],
-    where: `Trunk_Diameter > 10`
+    where: 'Trunk_Diameter > 10'
   }
   const results = winnow.query(features, options)
   t.equal(results.total_diameter, 735026)