+++ packages/winnow/src/test/fields.js
@@ -3,7 +3,7 @@ const test = require('tape')
 const fs = require('fs')
 const winnow = require('../')
 const path = require('path')
-const features = JSON.parse(fs.readFileSync(path.join(__dirname, 'fixtures', 'trees.min.geojson'))).features
+const features = JSON.parse(fs.readFileSync(path.join(__dirname, 'fixtures', 'trees.geojson'))).features
 
 test('Select a single field', (t) => {
   t.plan(1)